factoring out auth_github and delayed_interrupt, and using them in update-mathlib
