+++ scripts/update-mathlib.py
@@ -10,26 +10,8 @@ import urllib3
 import certifi
 import configparser
 import tarfile
-
-def auth_github():
-    try:
-        repo = Repo('.', search_parent_directories=True)
-        config = repo.config_reader()
-    except:
-        print('This does not seem to be a git repository.')
-        return Github()
-    try:
-        return Github(config.get('github', 'user'), config.get('github', 'password'))
-    except configparser.NoSectionError:
-        print('No github section found in \'git config\'')
-        return Github()
-    except configparser.NoOptionError:
-        try:
-            return Github(config.get('github', 'oauthtoken'))
-        except configparser.NoOptionError:
-            print('No github \'user\'/\'password\' or \'oauthtoken\' keys found in \'git config\'.')
-            print('You can create an OAuth token at https://github.com/settings/tokens/new (no scopes are required).')
-            return Github()
+from delayed_interrupt import DelayedInterrupt
+from auth_github import auth_github
 
 # find root of project and leanpkg.toml
 cwd = os.getcwd()
@@ -105,5 +87,6 @@ else:
 
 # Extract archive
 print("Extracting nightly...")
-ar = tarfile.open(os.path.join(mathlib_dir, asset.name))
-ar.extractall('_target/deps/mathlib')
+with DelayedInterrupt([signal.SIGTERM, signal.SIGINT]):
+	ar = tarfile.open(os.path.join(mathlib_dir, asset.name))
+	ar.extractall('_target/deps/mathlib')
